Feature/100 event generator

See merge request cash-flow-team/cash-flow!93

+++ client/lib/presentation/gameboard/game_events/debenture/ui/debenture_game_event.dart
@@ -37,7 +37,7 @@ class DebentureGameEvent extends HookWidget {
 
     final selectorViewModel = SelectorViewModel(
       currentPrice: eventData.currentPrice.toDouble(),
-      passiveIncomePerMonth: eventData.profitabilityPercent,
+      passiveIncomePerMonth: eventData.profitabilityPercent.toInt(),
       alreadyHave: alreadyHave,
       maxCount: eventData.availableCount,
       changeableType: true,
